Move `RedisCluster` to redis-test.
