+++ redis/benches/bench_cluster.rs
@@ -2,6 +2,7 @@
 #![cfg(feature = "cluster")]
 use criterion::{black_box, criterion_group, criterion_main, Criterion, Throughput};
 use redis::cluster::cluster_pipe;
+use redis_test::cluster::RedisClusterConfiguration;
 
 use support::*;
